+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented here.
 - Support connect in server and database
 
 ## [1.1.1] - 2025-08-20
+- Add retry connect to server
 - Add flag --all. Using this flag, you can dump all databases from the configuration.
 
 ## [1.1.0] - 2025-08-19